% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Assess Generalizability of Randomized Trial}
\usage{
assess(trial, selection_covariates, data, method = "lr", trim.pop = TRUE,
  is.data.disjoint = TRUE)
}
\arguments{
\item{data}{a data frame containing the variables specified in the model}

\item{method}{choose method to predict the probability of trial participation.  Default is logistic regression ("lr").  Other methods supported are random forests ("rf") and lasso ("lasso")}

\item{trim.pop}{logical, if TRUE - trim target population so that covariates do not exceed bounds of trial covariates}

\item{is.data.disjoint}{logical, if TRUE - trial and population data are considered independent.  If FALSE - trial is proper subset of population data.  See details for implications on weighting methods.}

\item{formula}{an object of class "formula". The formula specifying the model for trial participation.  Lefthand side should be a binary variable indicating trial membership, and righthand side should contain pre-treatment covariates measured in data set.}
}
\description{
Assess Generalizability of Randomized Trial
}
\examples{
assess(trial ~ age + sex + race, data = ctn_data)
}
