% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_index.R
\name{gen_index}
\alias{gen_index}
\title{Calculate `generalizability index` to describe how similar or different trial is from population}
\usage{
gen_index(dat1B, dat2B)
}
\arguments{
\item{dat1B}{vector of probabilities of trial participation among individuals in the trial}

\item{dat2B}{vector of probabilities of trial participation among individuals in the population}
}
\value{
the generalizability index, a value between 0 and 1, where scores greater than 1 indicate greater similarity (see Tipton paper for description)
}
\description{
Calculate `generalizability index` to describe how similar or different trial is from population
}
\examples{
gen_index(probs_population, probs_trial)
}
