% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_weights.R
\name{gen_weights}
\alias{gen_weights}
\title{Estimate weights for generalizing ATE by predicting probability of trial participation}
\usage{
gen_weights(trial, selection_covariates, data, method = "lr",
  is.data.disjoint = TRUE)
}
\arguments{
\item{data}{a data frame containing the variables specified in the model}

\item{method}{choose method to predict the probability of trial participation.  Default is logistic regression ("lr").  Other methods supported are random forests ("rf") and lasso ("lasso")}

\item{formula}{an object of class "formula". The formula specifying the model for trial participation.  Lefthand side should be a binary variable indicating trial membership, and righthand side should contain pre-treatment covariates measured in data set.}
}
\value{
\code{gen_weights} returns an object of the class "gen_weights", containing the following: \code{participation_probs} (predicted probabilities of trial participation), \code{weights} (weights constructed from predicted probabilities - see description for more details), \code{gen_index} (generalizability index)
}
\description{
Estimate weights for generalizing ATE by predicting probability of trial participation
}
\examples{
gen_weights(trial ~ age + sex + race, data = ctn_data)
gen_weights(trial ~ age + sex + race, data = ctn_data, method = 'rf')
}
