% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize_tmle.R
\name{generalize_tmle}
\alias{generalize_tmle}
\title{TMLE for estimating TATE}
\usage{
generalize_tmle(outcome, treatment, trial, selection_covariates, data,
  is_data_disjoint = TRUE, seed)
}
\arguments{
\item{outcome}{variable name denoting outcome}

\item{treatment}{variable name denoting binary treatment assignment (ok if only available in trial, not population)}

\item{trial}{variable name denoting binary trial participation (1 = trial participant, 0 = not trial participant)}

\item{selection_covariates}{vector of covariate names in data set that predict trial participation}

\item{data}{data frame comprised of "stacked" trial and target population data}

\item{is_data_disjoint}{logical. If TRUE, then trial and population data are considered independent.}

\item{seed}{numeric. By default, the seed is set to 13783, otherwise can be specified (such as for simulation purposes).}
}
\value{
\code{generalize_tmle} returns a list of the TATE estimate, standard error, and 95\% CI bounds
}
\description{
TMLE for estimating TATE
}
