% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_table.R
\name{covariate_table}
\alias{covariate_table}
\title{Create Covariate Balance Table}
\usage{
covariate_table(trial, selection_covariates, data, weighted_table = FALSE,
  selection_method = "lr", is_data_disjoint = TRUE)
}
\arguments{
\item{trial}{variable name denoting binary trial participation (1 = trial participant, 0 = not trial participant)}

\item{selection_covariates}{vector of covariate names in data set that predict trial participation}

\item{data}{data frame comprised of "stacked" trial and target population data}
}
\description{
Create Covariate Balance Table
}
