% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.R
\name{generalize}
\alias{generalize}
\title{Generalize Average Treatment Effect from Randomized Trial to Population}
\usage{
generalize(outcome, treatment, trial, selection_covariates, data,
  method = "weighting", selection_method = "lr",
  is_data_disjoint = TRUE, trim_pop = FALSE, seed)
}
\arguments{
\item{outcome}{variable name denoting outcome}

\item{treatment}{variable name denoting binary treatment assignment (ok if only available in trial, not population)}

\item{trial}{variable name denoting binary trial participation (1 = trial participant, 0 = not trial participant)}

\item{selection_covariates}{vector of covariate names in data set that predict trial participation}

\item{data}{data frame comprised of "stacked" trial and target population data}

\item{method}{method to generalize average treatment effect to the target population.  Default is "weighting" (weighting by participation probability).  Other methods supported are "BART" (Bayesian Additive Regression Trees - NOT READY YET) and "TMLE" (Targeted Maximum Likelihood Estimation)}

\item{selection_method}{method to estimate the probability of trial participation.  Default is logistic regression ("lr").  Other methods supported are Random Forests ("rf") and Lasso ("lasso")}

\item{is_data_disjoint}{logical. If TRUE, then trial and population data are considered independent.  This affects calculation of the weights - see details for more information.}

\item{trim_pop}{logical. If TRUE, then population data are subset to exclude individuals with covariates outside bounds of trial covariates.}

\item{seed}{numeric. By default, the seed is set to 13783, otherwise can be specified (such as for simulation purposes).}
}
\value{
\code{generalize} returns an object of the class "generalize"
}
\description{
Generalize Average Treatment Effect from Randomized Trial to Population
}
