% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{tmle}
\alias{tmle}
\title{TMLE}
\usage{
tmle(outcome, treatment, trial, selection_covariates, data)
}
\arguments{
\item{outcome}{variable denoting outcome}

\item{treatment}{variable denoting binary treatment assignment (ok if only available in trial, not population)}

\item{data}{a data frame containing the variables specified in the model}

\item{method}{choose method to generalize average treatment effect.  Default is "weighting" (weighting by the odds of participation probability).  Other methods supported are "bart" (Bayesian Additive Regression Trees - NOT READY YET) and "tmle" (Targeted Maximum Likelihood Estimation)}

\item{weight_method}{choose method to predict the probability of trial participation.  Default is logistic regression ("lr").  Other methods supported are random forests ("rf") and lasso ("lasso")}

\item{outcome_formula}{an object of class "formula." Can specify an optional outcome model to include pre-treatment covariates.}
}
\value{
\code{generalize} returns an object of the class "generalize", containing the following: \code{TATE} (target population average treatment effect), \code{TATE_CI} (95% Confidence Interval for TATE).  If outcome is binary, reports TATE as risk difference as well as odds ratio, with accompanying CIs
}
\description{
TMLE
}
<<<<<<< HEAD
=======
\examples{
generalize(outcome = "STUDYCOMPLETE", treatment = "treat", selection_formula = trial ~ age + sex + race, data = ctn_data, method = "weighting")
generalize(outcome = "STUDYCOMPLETE", treatment = "treat", selection_formula = trial ~ age + sex + race, data = ctn_data, method = "tmle")
}
>>>>>>> 9556e35d29cc067f2a5261ef6c3024119ac2d1d4
