% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_pop.R
\name{trim_pop}
\alias{trim_pop}
\title{Trim target population covariates to be within bounds of trial}
\usage{
trim_pop(trial, covariates, data, just_population = FALSE)
}
\arguments{
\item{trial}{variable name: must be binary indicator}

\item{covariates}{a vector of covariate names in data set that predict trial participation}

\item{data}{a data frame containing the variables specified in the model}

\item{just_population}{logical. if FALSE (default), function returns full data set, if TRUE, function only returns trimmed population data.}
}
\value{
\code{trim_pop} returns a data frame, where the target population covariates do not exceed the bounds of the trial covariates
}
\description{
Trim target population covariates to be within bounds of trial
}
\examples{
trim_pop(trial,covariates, data = ctn_data)
}
