% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_pop.R
\name{trim_pop}
\alias{trim_pop}
\title{Subset Population so Population Covariates are within bounds of Trial Covariates}
\usage{
trim_pop(trial, selection_covariates, data)
}
\arguments{
\item{trial}{variable name denoting binary trial participation (1 = trial participant, 0 = not trial participant)}

\item{selection_covariates}{vector of covariate names in data set that predict trial participation}

\item{data}{data frame comprised of "stacked" trial and target population data}
}
\value{
\code{trim_pop} returns a data frame, where the target population covariates do not exceed the bounds of the trial covariates
}
\description{
Subset Population so Population Covariates are within bounds of Trial Covariates
}
